+++ history.md
@@ -1,3 +1,7 @@
+# 2022-05-13
+
+- Notification support like VSCode https://github.com/neoclide/coc.nvim/discussions/3813
+
 # 2022-05-07
 
 - Show sort method as description in outline view.